Move functions from main to lib
* Makes main.rs cleaner and prepares for more tests in later commits.
